id changed from int to long

+++ library/src/main/java/com/heinrichreimersoftware/materialintro/slide/SimpleSlideActivity.java
@@ -4,6 +4,6 @@ import android.view.View;
 
 public interface SimpleSlideActivity
 {
-    void onSlideViewCreated(SimpleSlide.SimpleSlideFragment fragment, View view, int id);
-    void onSlideDestroyView(SimpleSlide.SimpleSlideFragment fragment, View view, int id);
+    void onSlideViewCreated(SimpleSlide.SimpleSlideFragment fragment, View view, long id);
+    void onSlideDestroyView(SimpleSlide.SimpleSlideFragment fragment, View view, long id);
 }
